requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiosqlite==0.17.0
    # via tortoise-orm
amqp==5.1.0
    # via kombu
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
bcrypt==3.2.0
    # via
    #   -r requirements.in
    #   passlib
billiard==3.6.4.0
    # via celery
celery==5.2.6
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==41.0.6
    # via
    #   pyjwt
    #   python-jose
dnspython==2.2.1
    # via email-validator
ecdsa==0.17.0
    # via python-jose
email-validator==1.1.3
    # via
    #   fastapi-users
    #   pydantic
fastapi==0.75.0
    # via
    #   -r requirements.in
    #   fastapi-users
fastapi-users[tortoise-orm]==9.3.0
    # via
    #   -r requirements.in
    #   fastapi-users-db-tortoise
fastapi-users-db-tortoise==2.0.0
    # via fastapi-users
gunicorn==20.1.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
iso8601==0.1.16
    # via tortoise-orm
jinja2==3.1.3
    # via -r requirements.in
kombu==5.2.4
    # via celery
makefun==1.13.1
    # via fastapi-users
markupsafe==2.1.1
    # via jinja2
paho-mqtt==2.1.0
    # via -r requirements.in
passlib[bcrypt]==1.7.4
    # via
    #   -r requirements.in
    #   fastapi-users
prompt-toolkit==3.0.29
    # via click-repl
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic[email]==1.9.0
    # via
    #   -r requirements.in
    #   fastapi
pyjwt[crypto]==2.3.0
    # via fastapi-users
pypika-tortoise==0.1.4
    # via tortoise-orm
python-dotenv==0.20.0
    # via -r requirements.in
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via fastapi-users
pytz==2022.1
    # via
    #   celery
    #   tortoise-orm
requests==2.27.1
    # via -r requirements.in
rsa==4.8
    # via python-jose
six==1.16.0
    # via
    #   bcrypt
    #   click-repl
    #   ecdsa
    #   python-multipart
sniffio==1.2.0
    # via anyio
starlette==0.17.1
    # via fastapi
tortoise-orm==0.18.1
    # via fastapi-users-db-tortoise
typing-extensions==4.1.1
    # via
    #   aiosqlite
    #   pydantic
urllib3==1.26.9
    # via requests
uvicorn==0.17.6
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
